Asm version 4 fixes in Jaqu gradle builder logging

+++ Jaqu/src/org/h2/jaqu/ext/common/CommonAssembly.java
@@ -29,6 +29,7 @@ import java.util.List;
 import org.h2.jaqu.ext.asm.JaquClassAdapter;
 import org.objectweb.asm.ClassReader;
 import org.objectweb.asm.ClassWriter;
+import org.objectweb.asm.Opcodes;
 
 /**
  * @author shai
@@ -36,46 +37,56 @@ import org.objectweb.asm.ClassWriter;
  */
 public class CommonAssembly {
 
-	public static BuildStats assenbleFiles(File outputDir, StringBuilder report) {
+	public static BuildStats assembleFiles(File outputDir, StringBuilder report) {
 		ArrayList<File> files = new ArrayList<File>();
 		getAllFiles(outputDir, files);
 		
-		int success = 0, failure = 0;
+		int success = 0, failure = 0, ignored = 0;
 		for (File classFile: files) {
 			try {
-				assembleFile(classFile);
-				report.append(String.format("SUCCESS -- %s\n", classFile));
-				success++;
+				if (assembleFile(classFile)) {
+					report.append(String.format("SUCCESS -- %s\n", classFile));
+					success++;
+				}
+				else
+					report.append(String.format("IGNORED -- %s\n", classFile));
+					ignored++;
 			}
 			catch (Exception e) {
 				report.append(String.format("FAILED -- %s --> %s\n", classFile, e.getMessage()));
 				failure++;
 			}
 		}
-		return new BuildStats(success, failure);
+		return new BuildStats(success, failure, ignored);
 	}
 	
-	private static void assembleFile(File classFile) throws Exception {
+	private static boolean assembleFile(File classFile) throws Exception {
 		FileInputStream fis = new FileInputStream(classFile);
         byte[] b = null;       
        
     	ClassReader cr = new ClassReader(fis);
         ClassWriter cw = new ClassWriter(ClassWriter.COMPUTE_MAXS + ClassWriter.COMPUTE_FRAMES);
             		
-        JaquClassAdapter jaquClassAdapter = new JaquClassAdapter(cw);
+        JaquClassAdapter jaquClassAdapter = new JaquClassAdapter(Opcodes.ASM4, cw);
         cr.accept(jaquClassAdapter, 0);
-            		
-        b = cw.toByteArray();
-        fis.close();
-        if (b != null) {
-        	classFile.delete();
-        	classFile.createNewFile();
-            FileOutputStream fos = new FileOutputStream(classFile);
-    		fos.write(b);
-    		fos.flush();
-    		fos.close();
+        
+        if (jaquClassAdapter.isJaquAnnotated()) {
+	        b = cw.toByteArray();
+	        fis.close();
+	        if (b != null) {
+	        	classFile.delete();
+	        	classFile.createNewFile();
+	            FileOutputStream fos = new FileOutputStream(classFile);
+	    		fos.write(b);
+	    		fos.flush();
+	    		fos.close();
+	        }
+	        return true;
         }
- 
+        else {
+        	fis.close();
+        	return false;
+        }
 	}
 
 	private static void getAllFiles(File outputDir, List<File> files){
